+++ src/components/admin-item-edit/admin-item-edit.component.tsx
@@ -1,0 +1,12 @@
+import AdminItemFormContainer from "../admin-item-form-container/admin-item-form-container.component";
+import { formTypes } from "../../types/types";
+
+const AdminItemEdit = () => {
+  return (
+    <div>
+      <AdminItemFormContainer formtype={formTypes["UPDATE"]} />;
+    </div>
+  );
+};
+
+export default AdminItemEdit;
