+++ src/routes/admin.component.tsx
@@ -3,6 +3,7 @@ import { Routes, Route } from "react-router-dom";
 import AdminDashboard from "../components/admin-dashboard/admin-dashboard.component";
 import AdminLogin from "../components/admin-login/admin-login.component";
 import AdminProfile from "../components/admin-profile/admin-profile.component";
+import AdminItemEdit from "../components/admin-item-edit/admin-item-edit.component";
 
 const Admin = () => {
   return (
@@ -10,6 +11,7 @@ const Admin = () => {
       <Routes>
         <Route index element={<AdminLogin />} />
         <Route path="dashboard" element={<AdminDashboard />} />
+        <Route path="dashboard/item/:id/edit" element={<AdminItemEdit />} />
         <Route path="profile" element={<AdminProfile />} />
       </Routes>
     </>
